fix: исправлены прото файлы
#25

+++ src/Server/Modules/Player/Module.Player.Api/PlayerExtensions.cs
@@ -11,7 +11,7 @@ internal static class PlayerExtensions
             Name = stats.Name,
             Health = stats.Health,
             Hunger = stats.Hunger,
-            PocketMoney = stats.PocketMoney,
+            PocketMoney = (double)stats.PocketMoney,
             Mood = stats.Mood,
         };
 }
